refactor: organize components in sub-folders per concept
