+++ src/components/staticPages/LandingPage/LandingPageComponent.style.ts
