+++ src/components/staticPages/LandingPage/index.ts
