+++ src/routes.ts
@@ -1,14 +1,14 @@
 import React from 'react'
 
-import LandingPage from './components/LandingPage'
-import CaseSetManager from './components/CaseSetManager'
-import CaseSetViewer from './components/CaseSetViewer'
-import CaseSetCreator from './components/CaseSetCreator'
-import AiImplementationManager from './components/AiImplementationManager'
-import BenchmarkRunner from './components/BenchmarkRunner'
-import BenchmarkCreator from './components/BenchmarkCreator'
-import BenchmarkEvaluator from './components/BenchmarkEvaluator'
-import BenchmarkingSessionManagerContainer from './components/BenchmarkingSessionManager'
+import LandingPage from './components/staticPages/LandingPage'
+import CaseSetManager from './components/caseSets/CaseSetManager'
+import CaseSetViewer from './components/caseSets/CaseSetViewer'
+import CaseSetCreator from './components/caseSets/CaseSetCreator'
+import AiImplementationManager from './components/aiImplementations/AiImplementationManager'
+import BenchmarkRunner from './components/benchmarkingSessions/BenchmarkRunner'
+import BenchmarkCreator from './components/benchmarkingSessions/BenchmarkCreator'
+import BenchmarkEvaluator from './components/benchmarkingSessions/BenchmarkEvaluator'
+import BenchmarkingSessionManagerContainer from './components/benchmarkingSessions/BenchmarkingSessionManager'
 
 const CASE_SETS_PATH = 'case-sets'
 const BENCHMARKING_SESSIONS_PATH = 'benchmarking-sessions'
